refactor(namecoin): split nip05.ts into focused submodules
Splits the 601-line src/lib/namecoin/nip05.ts into:

  - identifier.ts (shape validation + parsing)
  - transport.ts  (WSS / ElectrumX / name_show)
  - script.ts     (OP_NAME_UPDATE build + parse)
  - value.ts      (JSON value extraction)
  - hex.ts        (strict hex byte helpers)

nip05.ts now keeps only the high-level queryProfile / isValid
orchestrators plus a barrel re-export of the public surface, so
existing imports (cache.ts, profile/nip05.ts, api/nip05/verify,
the test file) continue to resolve unchanged.

Also tightens isValid to accept and forward a custom servers list,
so callers can pin the resolve + verify paths to the same server set.

Every new file lands well under the 420-line repository limit.
